Move to support for awssum circa this year
In order to have govcloud support we have to update awssum. I've
decided to move to the latest. Unfortunately ShareJS used to
have awssum as a soft requirement but it expected a certain
interface. I've now made it a hard requirement, since it is
for our use case, which made it easier to change the interface
to match how awssum works now.

This also includes being built by a newer version of CoffeeScript
so a bunch of checked-in built files also got changed. Whee.

The pertinent changes in this are in package.coff and amazon.coffee.
Everything else is generated